working on booking check admin page

+++ src/views/about/bookingAdmin/BookingButtonsAdmin.vue
@@ -12,7 +12,8 @@
         <b>
           {{ item.bookingTime }}시 {{ item.name }} ({{ item.phoneNumber }})</b
         >
-        <v-icon @click="deleteBooking(item.bookingDate, item.bookingTime)"
+        <v-icon
+          @click="deleteBooking(item.bookingDate, item.bookingTime, item.name)"
           >mdi-delete</v-icon
         >
       </v-chip>
@@ -25,7 +26,7 @@
 import firebase from "firebase/compat/app";
 import "firebase/compat/auth";
 import "firebase/compat/firestore";
-// import Swal from "sweetalert2";
+import Swal from "sweetalert2";
 export default {
   props: ["selectedDate"],
   data() {
@@ -46,17 +47,34 @@ export default {
   },
   methods: {
     // 예약리스트 삭제
-    deleteBooking(date, time) {
-      const db = firebase.firestore();
-      db.collection("booking")
-        .where("bookingDate", "==", date)
-        .where("bookingTime", "==", time)
-        .get()
-        .then((querySnapshot) => {
-          querySnapshot.forEach((doc) => {
-            doc.ref.delete();
-          });
-        });
+    deleteBooking(date, time, name) {
+      Swal.fire({
+        title: "예약삭제",
+        html: `${name}님의 ${date}<br>${time}시 예약을 정말 삭제하시겠습니까?`,
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "삭제",
+        cancelButtonText: "취소",
+      }).then((result) => {
+        if (result.isConfirmed) {
+          const db = firebase.firestore();
+          db.collection("booking")
+            .where("bookingDate", "==", date)
+            .where("bookingTime", "==", time)
+            .get()
+            .then((querySnapshot) => {
+              querySnapshot.forEach((doc) => {
+                doc.ref.delete();
+              });
+            });
+          Swal.fire("삭제완료", "예약이 삭제되었습니다.", "success");
+          setTimeout(() => {
+            location.reload();
+          }, 1500);
+        }
+      });
     },
     // 달력에서 선택된 날짜에 존재하는 예약들 가져오는 메서드
     getSelectedDate(paramSelectedDate) {
@@ -64,6 +82,7 @@ export default {
       const db = firebase.firestore();
       db.collection("booking")
         .where("bookingDate", "==", paramSelectedDate)
+        .orderBy("bookingTime")
         .get()
         .then((querySnapshot) => {
           querySnapshot.forEach((doc) => {
